게시글 삭제 기능 개선2

+++ src/main/java/com/novProject/service/comment/CommentService.java
@@ -43,7 +43,9 @@ public class CommentService {
     @Transactional
     public void delete(Long id){
        Comment comment = commentRepository.findByPosts_id(id);
+       if(comment != null){ //댓글이 없어도 삭제되도록
         commentRepository.delete(comment);
+       }
     }
 
     @Transactional
